Consolidate JWKS Cache tests
